(hotfix) lock in rack dependency for now
This is needed to ensure the test suite passes in modern ruby
versions as well as 1.8.7.

We still have a few legacy apps running 1.8.7 that need this and
I want to make sure this gem will run for them for now.  Eventually
we will drop this requirement and test this gem again modern Rack.
But for now I need to continue locking in a 1.8.7 compatible Rack
for testing purposes.